Update README.md
#README. md

##SMART-ETHER-DATA-EMULATE-REPOSITORY

## Welcome to GitHub Pages

You can use the [editor on GitHub](https://github.com/https://github/ruzyysmartt/Smart-ether-data-emulate-repository.gits) to maintain and preview the content for your website in Markdown files.

Whenever you commit to this repository, GitHub Pages will run [Jekyll](https://jekyllrb.com/ruzyysmartt/Smart-ether-data-emulate-repository) to rebuild the pages in your site, from the content in your Markdown files.


### Markdown

Markdown is a lightweight and easy-to-use syntax for styling your writing. It includes conventions for

```markdown

Syntax highlighted code block

# Header 1

## Header 2

### Header 3

- Bulleted

- List

1. Numbered

2. List

**Bold** and _Italic_ and `Code` text

[Link](url) and ![Image](src)

```

For more details see [GitHub Flavored Markdown](https://guides.github.com/features/mastering-markdown/).

### Jekyll Themes

Your Pages site will use the layout and styles from the Jekyll theme you have selected in your [repository settings](https://github.com/ruzyysmartt/Ruzyysmartt-ether-Web3/settings). The name of this theme is saved in the Jekyll `_config.yml` configuration file.

### Support or Contact

Having trouble with Pages? Check out our [do0'fcccumentation](https://help.github.com/categories/github-pages-basics/) or [contact o](https://github.com/ruzyysmartt/contact) and we’ll help you sort it out.



## WHITECONTRACT VERlFY
Etherscan.io 18/01/2020 19:03:45] I, hereby verify the owner/of the address [0x62991A1B4187481C8B5BB49Fa567711e09fF488D]

##License

https://github.com/ruzyysmartt/Smart-ether-data-emulate-repository.gits

Copyright (c) 2020 eveem-org

Permission is hereby granted, free of charge, to any person obtaining a copy

of this software and associated documentation files (the "Software"), to deal

in the Software without restriction, including without limitation the rights

to use, copy, modify, merge, publish, distribute, sublicense, and/or sell

copies of the Software, and to permit persons to whom the Software is

furnished to do so, subject to the following conditions:

The above copyright notice and tuhis permission notice shall be included in all

copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR

IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,

FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE

AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER

LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,

OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE

SOFTWARE.

more info:

https://github.com/ruzyysmartt/Smart-ether-data-emulate-repository.gits
https://github.com/https-github-com-ruzyysmartt/panoramix.git

©️GitHub.2020